+++ harvey/__init__.py
@@ -4,7 +4,7 @@ from harvey.containers import Container
 from harvey.deployments import Deployment
 from harvey.git import Git
 from harvey.messages import Message
-from harvey.utils import Utils
+from harvey.utils.utils import Utils
 from harvey.webhooks import Webhook