+++ harvey/deployments.py
@@ -14,9 +14,13 @@ import yaml
 from harvey.config import Config
 from harvey.containers import Container
 from harvey.git import Git
-from harvey.locks import Lock
 from harvey.messages import Message
-from harvey.utils import Utils
+from harvey.repos.deployments import store_deployment_details
+from harvey.repos.locks import (
+    lookup_project_lock,
+    update_project_lock,
+)
+from harvey.utils.utils import Utils
 from harvey.webhooks import Webhook
 
 
@@ -30,7 +34,7 @@ class Deployment:
 
         try:
             # Kill the deployment if the project is locked
-            if Lock.lookup_project_lock(Webhook.repo_full_name(webhook))['locked'] is True:
+            if lookup_project_lock(Webhook.repo_full_name(webhook))['locked'] is True:
                 Utils.kill_deployment(
                     (
                         f'{Webhook.repo_full_name(webhook)} deployments are locked. Please try again later or unlock'
@@ -47,12 +51,12 @@ class Deployment:
 
         start_time = datetime.datetime.utcnow()
 
-        _ = Lock.update_project_lock(
+        _ = update_project_lock(
             project_name=Webhook.repo_full_name(webhook),
             locked=True,
             system_lock=True,
         )
-        Utils.store_deployment_details(webhook)
+        store_deployment_details(webhook)
         # Run git operation first to ensure the config is present and up-to-date
         git = Git.update_git_repo(webhook)
 
@@ -104,56 +108,64 @@ class Deployment:
         """After receiving a webhook, spin up a deployment based on the config.
         If a Deployment fails, it fails early in the individual functions being called.
         """
-        logger = woodchips.get(Config.logger_name)
+        try:
+            logger = woodchips.get(Config.logger_name)
 
-        webhook_config, webhook_output, start_time = Deployment.initialize_deployment(webhook)
-        deployment = webhook_config.get('deployment_type', Config.default_deployment).lower()
+            webhook_config, webhook_output, start_time = Deployment.initialize_deployment(webhook)
+            deployment = webhook_config.get('deployment_type', Config.default_deployment).lower()
 
-        if deployment == 'deploy':
-            deploy_output = Deployment.deploy(webhook_config, webhook, webhook_output)
+            if deployment == 'deploy':
+                deploy_output = Deployment.deploy(webhook_config, webhook, webhook_output)
 
-            healthcheck = webhook_config.get('healthcheck')
-            healthcheck_messages = 'Healthchecks:\n'
-            docker_client = Container.create_client()
+                healthcheck = webhook_config.get('healthcheck')
+                healthcheck_messages = 'Healthchecks:\n'
+                docker_client = Container.create_client()
 
-            if healthcheck:
-                container_healthcheck_statuses = {}
-                for container in healthcheck:
-                    container_healthcheck = Container.run_container_healthcheck(docker_client, container, webhook)
-                    container_healthcheck_statuses[container] = container_healthcheck
-                    if container_healthcheck is True:
-                        healthcheck_message = f'\n{container} Healthcheck: {Message.success_emoji}'
-                    else:
-                        healthcheck_message = f'\n{container} Healthcheck: {Message.failure_emoji}'
-                    healthcheck_messages += healthcheck_message
+                if healthcheck:
+                    container_healthcheck_statuses = {}
+                    for container in healthcheck:
+                        container_healthcheck = Container.run_container_healthcheck(docker_client, container, webhook)
+                        container_healthcheck_statuses[container] = container_healthcheck
+                        if container_healthcheck is True:
+                            healthcheck_message = f'\n{container} Healthcheck: {Message.success_emoji}'
+                        else:
+                            healthcheck_message = f'\n{container} Healthcheck: {Message.failure_emoji}'
+                        healthcheck_messages += healthcheck_message
 
-                healthcheck_values = container_healthcheck_statuses.values()
-                all_healthchecks_passed = any(healthcheck_values) and list(healthcheck_values)[0] is True
-            else:
-                all_healthchecks_passed = True  # Set to true here since we cannot determine, won't kill the deploy
+                    healthcheck_values = container_healthcheck_statuses.values()
+                    all_healthchecks_passed = any(healthcheck_values) and list(healthcheck_values)[0] is True
+                else:
+                    all_healthchecks_passed = True  # Set to true here since we cannot determine, won't kill the deploy
 
-            end_time = datetime.datetime.utcnow()
-            execution_time = f'Deployment execution time: {end_time - start_time}'
-            logger.debug(f'{Webhook.repo_full_name(webhook)} {execution_time}')
-            final_output = f'{webhook_output}\n{deploy_output}\n{execution_time}\n{healthcheck_messages}\n'
+                end_time = datetime.datetime.utcnow()
+                execution_time = f'Deployment execution time: {end_time - start_time}'
+                logger.debug(f'{Webhook.repo_full_name(webhook)} {execution_time}')
+                final_output = f'{webhook_output}\n{deploy_output}\n{execution_time}\n{healthcheck_messages}\n'
 
-            if all_healthchecks_passed or not healthcheck:
+                if all_healthchecks_passed or not healthcheck:
+                    Utils.succeed_deployment(final_output, webhook)
+                else:
+                    Utils.kill_deployment(
+                        message=final_output,
+                        webhook=webhook,
+                    )
+            elif deployment == 'pull':
+                # We simply assign the final message because if we got this far, the repo has already been pulled
+                pull_success_message = (
+                    f'Harvey pulled {Webhook.repo_full_name(webhook)} successfully. {Message.success_emoji}\n'
+                )
+                logger.info(pull_success_message)
+                final_output = f'{webhook_output}\n{pull_success_message}'
                 Utils.succeed_deployment(final_output, webhook)
             else:
                 Utils.kill_deployment(
-                    message=final_output,
-                    webhook=webhook,
+                    f'deployment_type invalid, must be one of {Config.supported_deployments}', webhook
                 )
-        elif deployment == 'pull':
-            # We simply assign the final message because if we got this far, the repo has already been pulled
-            pull_success_message = (
-                f'Harvey pulled {Webhook.repo_full_name(webhook)} successfully. {Message.success_emoji}\n'
-            )
-            logger.info(pull_success_message)
-            final_output = f'{webhook_output}\n{pull_success_message}'
-            Utils.succeed_deployment(final_output, webhook)
-        else:
-            Utils.kill_deployment(f'deployment_type invalid, must be one of {Config.supported_deployments}', webhook)
+        except Exception as error:
+            # We wrap this entire block in a try/catch in an attempt to catch anything that bubbles to the
+            # top before hitting sentry as this function is the top-level function called when a thread has
+            # been spawned.
+            Utils.kill_deployment(f'Deployment failed: {error}.', webhook)
 
     @staticmethod
     def open_project_config(webhook: Dict[str, Any]):